Add support for StorageCarts and Powered MineCarts in Dispenser chests
Unfortunately, the contents of the StorageCart still wind up being dumped
on the ground. Every way I try getting access to the contents comes up with
an array of nulls. There are also some commented out logging lines from my
debugging attempts. I'm committing as is in case someone wants to give it
a poke.