pipes: no longer handling jukeboxes (postponed fix)
Jukeboxes are handled in a way that seems very much nonsensical: records are only put into
the box if there's already one playing (which is then >lost<), while records that are sucked
into the pipe are not removed from the player, allowing to duplicate items. Since Jukeboxes
are of low priority and but a gimmic, their fix is hereby postponed.