pipes: now adding the first block in #enumeratePipeBlocks to the visited-set
While using my custom pipe-search addon I noticed how the first block has always been
enumerated twice - this is due to the fact that it is added ahead of the loop and then, while
enumerating, the algorithm will walk "back" into this block when at the first successor
relative to the origin. Despite it being irrelevant for pipes themselves, seeing how they
ignore sticky pistons, this is definitely the more correct approach to handle matters.